[refactor] order

+++ Intro-Vulkan/VulkanApp/Main.cpp
@@ -1,11 +1,10 @@
 #define GLFW_INCLUDE_VULKAN
 #define GLM_FORCE_DEPTH_ZERO_TO_ONE
 
-#include <GLFW/glfw3.h>
 #include <stdexcept>
-#include <vector>
-
 #include <string>
+#include <vector>
+#include <GLFW/glfw3.h>
 using std::string;
 
 #include "VulkanRenderer.h"
@@ -43,11 +42,11 @@ int main()
 	{
 		glfwPollEvents();
 
-		float now = glfwGetTime();
+		float now = (float)glfwGetTime();
 		deltaTime = now - lastTime;
 		lastTime = now;
 
-		angle += 10.0 * deltaTime;
+		angle += 10.0f * deltaTime;
 		if (angle > 360.0f) { angle -= 360.0f; }
 
 		glm::mat4 firstModel(1.0f);
